1st draft of mvp refactor
