+++ app/src/androidTest/java/com/example/photogalleryapp/ExampleInstrumentedTest.java
@@ -21,6 +21,8 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
+import com.example.photogalleryapp.view.MainActivity;
+
 import org.hamcrest.Matchers;
 import org.junit.FixMethodOrder;
 import org.junit.Rule;
