+++ app/src/main/AndroidManifest.xml
@@ -20,11 +20,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.PhotoGalleryApp">
        <activity
-            android:name=".SearchActivity"
+            android:name=".view.SearchActivity"
             android:exported="true"
-            android:parentActivityName=".MainActivity" />
+            android:parentActivityName=".view.MainActivity" />
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -45,7 +45,7 @@
 
         <service
             android:foregroundServiceType="location"
-            android:name="com.example.photogalleryapp.MainActivity" />
+            android:name="com.example.photogalleryapp.view.MainActivity" />
 
     </application>
